Updated the Mammoth ItemLocale Controller to send an Item's Vendor information and the OrderedByPredictix flag to Mammoth
Updated the Mammoth Web API to store Item Vendor information in the ItemLocale_Supplier_XX tables and the OrderedByPredictix flag in the ItemAttributes_Locale_XX table
Added staging table for the Vendor/Supplier information updates in Mammoth
Updated unit tests

Related work items: #24440, #24859

+++ Icon/Mammoth/MammothDb/Mammoth.Db/dbo/Tables/ItemLocale_Supplier_NC.sql
@@ -6,7 +6,7 @@
     BusinessUnitID          INT             NOT NULL,
     SupplierName            NVARCHAR(255)   NOT NULL,
     SupplierItemID          NVARCHAR(20)    NULL,
-    SupplierCaseSize        INT             NULL,
+    SupplierCaseSize		DECIMAL(9,4)	NULL,
     IrmaVendorKey           NVARCHAR(10)    NULL,
     AddedDateUtc            DATETIME2(7)    DEFAULT SYSUTCDATETIME() NOT NULL,
     ModifiedDateUtc         DATETIME2(7)    NULL,
